Configuration sketch 99 percent done
For the most part, this is gtg, but I might like to add in some checks to see if the device has already been configured for SFE standards.

This check could play into user prompts, and make it slightly less confusing if you plug in a pre-configured device.

Added statuses for locking of config zone and data/otp zones. This makes the examples slightly cleaner - by just looking at booleans, and not showing the actual config bytes.

Adjusted formatting on the publicKey printing (in generatePublicKey), so that you can easily copy from serial terminal into another sketch. Hell yeah for modulus operator wahoo!

removed a lot of default debug.

Added public arrays for serialNumber and RevisionNumber. This makes example code a bit cleaner. All is updated in readConfigZone().