Zero-ing out example arrays - removed generatePublicKey at printInfo
Zero-ing out example arrays for key, message, and signature. This will hopefully highlight the fact that you need to copy/paste from example 2.

removing the "startup" print of the devices public key will also help highlight that we are using an EXTERNAL public key, and hopefully clarifies that this is different than the public key we saw in configuration print outs.