Example6_Challenge Alice and Bob sketches working.
Note these at now at 115200 for debug and serial... probably going to want to change the other examples to be consistent.

Also found a better way to wait for incoming message between alice and bob. My timeout while loops also use Serial1.available() and wait for the proper amount of bytes to come in. Then we know that the entire message is done transmitting and won't miss anything.